requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0
MarkupSafe==3.0.3
pycparser==2.23
python-decouple==3.8
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.5
tzdata==2025.3
Werkzeug==3.1.4
